+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Change Log
 Unreleased
 **********
 
+* Add author view to the XBlock to be used in Studio.
+
 0.8.4 - 2023-10-02
 **********************************************